load preference from mod
